LangGraph/project1/requirements.txt
langchain==0.3.25
langgraph==0.2.60
langchain-openai==0.2.14
openai==1.58.1
httpx==0.27.2
pydantic==2.10.4
typing-extensions==4.12.2
python-dotenv==1.0.1
orjson==3.10.12
ipython==8.31.0
matplotlib==3.10.0
networkx==3.4.2
pillow==10.4.0
ipykernel==7.2.0
langgraph-checkpoint-sqlite==2.0.0
langchain-tavily==0.1.6
langgraph-cli[inmem]==0.1.80
